+++ packages/utils-shared/src/isRequestWithoutBody.ts
@@ -1,0 +1,8 @@
+import { RequestMethod } from 'fatcher';
+
+/**
+ * Check this request method whether has request body.
+ */
+export function isRequestWithoutBody(method: RequestMethod) {
+    return ['GET', 'HEAD'].includes(method);
+}